1) made it so that training set and testing set are split based on time. 2) Changed the actual result to be the correct actual result to compare with
